+++ src/PeerSharp/GlobalUsings.cs
@@ -2,4 +2,5 @@ global using PeerSharp.Config;
 global using PeerSharp.Exceptions;
 global using PeerSharp.Interfaces;
 global using PeerSharp.Core;
+global using System.Globalization;